framer/requirements.txt
aiokafka
minio
python-dotenv
pillow
opencv-python